# BUG-149: Fix compatibilities performance bug