Use fetcher instead of new when we use Gem::SpecFetcher

Fixes: https://github.com/fastruby/next_rails/issues/149

## Description
<!--- Describe your changes in detail -->

Use [fetcher](https://rubyapi.org/3.4/o/gem/specfetcher#method-c-fetcher) instead of new when we use Gem::SpecFetcher
Fixes: https://github.com/fastruby/next_rails/issues/149


## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
<!--- Include any relevant details about your testing environment and the steps you followed -->
<!--- For example: I am using [Safari|Firefox|Chrome] then I visit [the users path] -->

Using `time` before and after:
- Before: `42.34s user 1.83s system 97% cpu 45.095 total`
- After: `5.68s user 0.61s system 72% cpu 8.728 total`

Using automated tests.

## Screenshots:
<!-- Add screenshots (applicable to any UI changes) -->

Using stack prof:
| Before | After |
| - | - |
| ![image](https://github.com/user-attachments/assets/c1ea2e8f-1289-4588-b1cd-6b830164d6b5)  |  ![image](https://github.com/user-attachments/assets/eba61493-fc91-49eb-90b9-99fbb801c81e)  |

**I will abide by the [code of conduct](https://github.com/fastruby/next_rails/blob/main/CODE_OF_CONDUCT.md)**